requirements.txt
langchain==0.3.20
langchain-community==0.3.19
langchain-google-genai==2.0.7
langchain-openai==0.3.8
python-dotenv==1.0.1
pypdf==5.3.1
faiss-cpu